+++ ml/src/test/java/org/apache/hama/ml/regression/LinearRegressionModelTest.java
@@ -17,11 +17,13 @@
  */
 package org.apache.hama.ml.regression;
 
+import static org.junit.Assert.assertEquals;
+
+import java.math.BigDecimal;
+
 import org.apache.hama.ml.math.DenseDoubleVector;
 import org.apache.hama.ml.math.DoubleVector;
 import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Testcase for {@link LinearRegressionModel}
@@ -34,15 +36,15 @@ public class LinearRegressionModelTest {
     DoubleVector x = new DenseDoubleVector(new double[]{2, 3, 4});
     double y = 1;
     DoubleVector theta = new DenseDoubleVector(new double[]{1, 1, 1});
-    Double cost = linearRegressionModel.calculateCostForItem(x, y, 2, theta);
-    assertEquals("wrong cost calculation for linear regression", Double.valueOf(16d), cost);
+    BigDecimal cost = linearRegressionModel.calculateCostForItem(x, y, 2, theta);
+    assertEquals("wrong cost calculation for linear regression", BigDecimal.valueOf(16d), cost);
   }
 
   @Test
   public void testCorrectHypothesisCalculation() throws Exception {
     LinearRegressionModel linearRegressionModel = new LinearRegressionModel();
-    Double hypothesisValue = linearRegressionModel.applyHypothesis(new DenseDoubleVector(new double[]{1, 1, 1}),
+    BigDecimal hypothesisValue = linearRegressionModel.applyHypothesis(new DenseDoubleVector(new double[]{1, 1, 1}),
             new DenseDoubleVector(new double[]{2, 3, 4}));
-    assertEquals("wrong hypothesis value for linear regression", Double.valueOf(9), hypothesisValue);
+    assertEquals("wrong hypothesis value for linear regression", BigDecimal.valueOf(9d), hypothesisValue);
   }
 }
